LayoutTests/http/tests/webfont/fontloader-loading-attribute-expected.txt

Test for FontLoader.loading attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.fonts.checkFont('10px TestFont') is false
PASS document.fonts.loading is false
PASS document.fonts.loading is true
PASS document.fonts.checkFont('10px TestFont') is true
PASS document.fonts.loading is false
PASS successfullyParsed is true

TEST COMPLETE
